+++ RELEASES.rst
@@ -790,6 +790,10 @@ Planned
   body was not checked and Duktape would accept e.g. "function () [}"
   (see GH-106)
 
+* Fix compiler register limit bug where an Ecmascript function with
+  [65536,262143] VM registers would be compiled into incorrect bytecode
+  instead of being rejected with an internal error (see GH-111)
+
 2.0.0 (XXXX-XX-XX)
 ------------------